Test Database
I have converted our ER diagram into a relational database model. There isn't any data yet but I managed to successfully link the tables using the agreed upon primary/foreign keys. 

All you have to do is download this sql file and import it in myPHPAdmin (XAMPP). Then you can head over to the designer tab (under 'more'), to see the visualisation of our database.

Please note that this is only a test. We need first check if our ER diagram was fine and we also need to double check if I've normalised the table correctly. This was made to experiment